patronage_2019_cognitive_task_2/requirements.txt
apsw==3.16.2.post1
arandr==0.1.9
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
beautifulsoup4==4.6.0
bottle==0.12.13
certifi==2018.10.15
chardet==3.0.4
CherryPy==8.9.1
click==5.1
colorama==0.4.0
cryptography==2.1.4
cssselect==1.0.3
cssutils==1.0.2
cycler==0.10.0
dj-database-url==0.3.0
dj-static==0.0.6
Django==1.11.16
django-appconf==0.6
django-compressor==1.4
djangorestframework==3.0.0
dnspython==1.15.0
drf-nested-routers==0.9.0
enum34==1.1.6
f.lux-indicator-applet==1.1.11rc0
feedparser==5.2.1
gunicorn==19.1.1
gyp==0.1
html5-parser==0.4.4
html5lib==0.999999999
idna==2.7
image==1.5.27
ipaddress==1.0.17
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.0.1
lxml==4.2.1
Markdown==2.6.9
MAT==0.6.1
matplotlib==2.2.3
mechanize==0.2.5
msgpack==0.5.6
mutagen==1.38
netifaces==0.10.4
numpy==1.15.4
olefile==0.45.1
pandas==0.23.4
pbr==3.1.1
pdfrw==0.4
pexpect==4.2.1
Pillow==5.3.0
platformio==3.6.0
pycairo==1.16.2
pycrypto==2.6.1
Pygments==2.2.0
pygobject==3.26.1
pyOpenSSL==17.5.0
pyparsing==2.3.0
pyserial==3.4
python-dateutil==2.7.5
pytz==2018.7
pyxdg==0.25
PyYAML==3.12
regex==2017.12.12
repoze.lru==0.7
requests==2.19.1
Routes==2.4.1
scikit-learn==0.20.1
scipy==1.2.0
SecretStorage==2.3.1
semantic-version==2.6.0
simplejson==3.13.2
six==1.12.0
sklearn==0.0
sqlparse==0.2.4
static3==0.5.1
stevedore==1.28.0
subprocess32==3.5.3
urllib3==1.23
uTidylib==0.3
virtualenv==15.1.0
virtualenv-clone==0.2.5
virtualenvwrapper==4.3.1
webencodings==0.5
WebOb==1.7.3
